+++ Object_detection_image.py
@@ -1,40 +1,3 @@
-# product_details={
-#     1: {'id': 1, 'name': 'coke-1.5l', 'price': 72},
-#     2: {'id': 2, 'name': 'coke-700ml', 'price': 79},
-#     3: {'id': 3, 'name': 'coke-bottle-2l', 'price': 89},
-#     4: {'id': 4, 'name': 'coke-bottle-2l-zero', 'price': 101},
-#     5: {'id': 5, 'name': 'coke-bottle-2ll', 'price': 120},
-#     6: {'id': 6, 'name': 'coke-bottle-300ml', 'price': 30},
-#     7: {'id': 7, 'name': 'coke-bottle-700ml', 'price': 40},
-#     8: {'id': 8, 'name': 'coke-can', 'price': 15},
-#     9: {'id': 9, 'name': 'coke-can-classic', 'price': 10},
-#     10: {'id': 10, 'name': 'coke-can-diet', 'price': 25},
-#     11: {'id': 11, 'name': 'coke-can-energy', 'price': 25},
-#     12: {'id': 12, 'name': 'coke-can-espresso', 'price': 26},
-#     13: {'id': 13, 'name': 'coke-can-orange', 'price': 20},
-#     14: {'id': 14, 'name': 'coke-can-zero', 'price': 29},
-#     15: {'id': 15, 'name': 'coke-glass', 'price': 10},
-#     16: {'id': 16, 'name': 'coke-glass-diet', 'price': 15},
-#     17: {'id': 17, 'name': 'coke-light-bottle-300ml', 'price': 30},
-#     18: {'id': 18, 'name': 'coke-zero-bottle-2l', 'price': 101},
-#     19: {'id': 19, 'name': 'coke-zero-bottle-700ml', 'price': 40},
-#     20: {'id': 20, 'name': 'gooday-large', 'price': 150},
-#     21: {'id': 21, 'name': 'gooday-medium', 'price': 60},
-#     22: {'id': 22, 'name': 'gooday-mediuml-choco', 'price': 70},
-#     23: {'id': 23, 'name': 'gooday-small', 'price': 20},
-#     24: {'id': 24, 'name': 'gooday-small-5', 'price': 5},
-#     25: {'id': 25, 'name': 'gooday-small-butter', 'price': 10},
-#     26: {'id': 26, 'name': 'gooday-small-choco', 'price': 10},
-#     27: {'id': 27, 'name': 'lays-large', 'price': 40},
-#     28: {'id': 28, 'name': 'lays-med', 'price': 20},
-#     29: {'id': 29, 'name': 'lays-small', 'price': 10},
-#     30: {'id': 30, 'name': 'lays-small-5', 'price': 5},
-#     31: {'id': 31, 'name': 'pepsi-can', 'price': 15},
-#     32: {'id': 32, 'name': 'pepsi-can-0', 'price': 20},
-#     33: {'id': 33, 'name': 'pepsi-can-pack-2', 'price': 40},
-#     34: {'id': 34, 'name': 'sprite-can', 'price': 15},
-#     35: {'id': 35, 'name': 'thumbsup-glass', 'price': 10}
-# }
 product_details={
     1: {
     'id': 1,
@@ -432,26 +395,6 @@ def perform_product_detection(PATH_TO_IMAGE):
     printcount =0;
 
     if count!=0:
-        # Draw the results of the detection (aka 'visulaize the results')
-
-        # vis_util.visualize_boxes_and_labels_on_image_array(
-        #     image,
-        #     np.squeeze(boxes),
-        #     np.squeeze(classes).astype(np.int32),
-        #     np.squeeze(scores),
-        #     category_index,
-        #     use_normalized_coordinates=True,
-        #     line_thickness=3,
-        #     min_score_thresh=0.60)
-
-        # # # All the results have been drawn on image. Now display the image.
-        # cv2.imshow('Object detector', image)
-
-        # # # Press any key to close the image
-        # cv2.waitKey(0)
-
-        # # # Clean up
-        # cv2.destroyAllWindows()
         for i in classes[0]:
             if(printcount == count):
                 break
@@ -461,6 +404,6 @@ def perform_product_detection(PATH_TO_IMAGE):
                 product_writer = csv.writer(product_file)
                 lister.append(category_index[i]['id'])
                 product_writer.writerow(lister)
-        return 1
+        return [1,count]
     else:
-        return PATH_TO_IMAGE
+        return [PATH_TO_IMAGE,count]
