+++ product_list.csv
@@ -1,1 +1,3 @@
-32
+16
+16
+6
