+++ requirements.txt
@@ -17,6 +17,7 @@ Flask==1.1.1
 Flask-Jsonpify==1.5.0
 Flask-RESTful==0.3.7
 Flask-SQLAlchemy==2.4.1
+Flask-Uploads==0.2.1
 gast==0.3.2
 google-pasta==0.1.7
 grpcio==1.23.0
@@ -51,7 +52,7 @@ pandas==0.25.1
 pandocfilters==1.4.2
 parso==0.5.1
 pickleshare==0.7.5
-Pillow==6.1.0
+Pillow==6.2.0
 prometheus-client==0.7.1
 prompt-toolkit==2.0.9
 protobuf==3.9.1
