+++ templates/InventoryPage.html
@@ -27,7 +27,7 @@
 
             <div class="info-box-content">
               <span class="info-box-text">Sales</span>
-              <span class="info-box-number">760</span>
+              <span class="info-box-number">{{ sales }}</span>
             </div>
             <!-- /.info-box-content -->
           </div>
