+++ templates/mainLayout.html
@@ -47,9 +47,7 @@
 <script type="text/javascript" src="{{ url_for('static',filename='js/dashboard2.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static',filename='js/adminlte.min.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static',filename='js/Chart.min.js') }}"></script>
-<script type="text/javascript" src="{{ url_for('static',filename='js/popper.min.js') }}"></script>
-<script type="text/javascript" src="{{ url_for('static',filename='js/gvsScript.js') }}"></script>
-
+<script type="text/javascript" src="{{ url_for('static',filename='js/customScript.js') }}"></script>
 </body>
 </html>
